added policies for album categories and albums, toastify and fixed sidebar

+++ app/Http/Controllers/Admin/AlbumCategoryController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\EditAlbumCategoryRequest;
 use App\Models\AlbumCategories;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class AlbumCategoryController extends \App\Http\Controllers\Controller
@@ -29,7 +30,7 @@ class AlbumCategoryController extends \App\Http\Controllers\Controller
         $albumsCategory->user_id = Auth::id();
         $res = $albumsCategory->save();
 
-        $messaggio = $res ? 'Categoria ' . $albumsCategory->id . ' inserita correttamente' : 'Categoria non inserita';
+        $messaggio = $res ? 'Categoria ID: ' . $albumsCategory->id . ' inserita correttamente' : 'Categoria non inserita';
         $tipoMessaggio = $res ? 'success' : 'danger';
         session()->flash('message', ['tipo' => $tipoMessaggio, 'testo' => $messaggio]);
 
@@ -48,6 +49,11 @@ class AlbumCategoryController extends \App\Http\Controllers\Controller
 
     public function update(EditAlbumCategoryRequest $request, AlbumCategories $category)
     {
+        $response = Gate::inspect('update', $category);
+        if ($response->denied()) {
+            session()->flash('message', ['tipo' => 'danger', 'testo' => 'Non hai i permessi per modificare questa categoria']);
+            return redirect()->back();
+        }
         $oldCategory = $category->category_name;
         $category->category_name = $request->input('category_name');
         if ($oldCategory != $category->category_name) {
@@ -56,7 +62,7 @@ class AlbumCategoryController extends \App\Http\Controllers\Controller
             $res = 0;
         }
 
-        $messaggio = $res ? 'Categoria ID : ' . $category->id . ' - Aggiornata correttamente' : 'Categoria ID : ' . $category->id . ' - Non aggiornata';
+        $messaggio = $res ? 'Categoria ID: ' . $category->id . ' - Aggiornata correttamente' : 'Categoria ID: ' . $category->id . ' - Non aggiornata';
         $tipoMessaggio = $res ? 'success' : 'danger';
         session()->flash('message', ['tipo' => $tipoMessaggio, 'testo' => $messaggio]);
 
@@ -65,7 +71,17 @@ class AlbumCategoryController extends \App\Http\Controllers\Controller
 
     public function destroy(AlbumCategories $category)
     {
+        $response = Gate::inspect('delete', $category);
+        if ($response->denied()) {
+            session()->flash('message', ['tipo' => 'danger', 'testo' => 'Non hai i permessi per eliminare questa categoria']);
+            return redirect()->back();
+        }
+
         $res = $category->delete();
+        $messaggio = $res ? 'Categoria ID: ' . $category->id . ' - Eliminata correttamente' : 'Categoria ID: ' . $category->id . ' - Non eliminata';
+        $tipoMessaggio = $res ? 'success' : 'danger';
+        session()->flash('message', ['tipo' => $tipoMessaggio, 'testo' => $messaggio]);
+        return redirect()->route('album.categories.index');
     }
 
     public function destroyBatch(Request $request)
@@ -74,6 +90,19 @@ class AlbumCategoryController extends \App\Http\Controllers\Controller
         if ($recordIds == null) {
             return;
         }
+        $categories = [];
+        foreach ($recordIds as $id) {
+            $categories[] = AlbumCategories::find($id);
+        }
+        $response = Gate::inspect('delete', $categories);
+        if ($response->denied()) {
+            session()->flash('message', ['tipo' => 'danger', 'testo' => 'Non hai i permessi per eliminare le categorie selezionate']);
+            return redirect()->back();
+        }
         $res = AlbumCategories::whereIn('id', $recordIds)->delete();
+        $messaggio = $res ? 'Categorie eliminate correttamente' : 'Categorie non eliminate';
+        $tipoMessaggio = $res ? 'success' : 'danger';
+        session()->flash('message', ['tipo' => $tipoMessaggio, 'testo' => $messaggio]);
+        return redirect()->route('album.categories.index');
     }
 }
